Move duckdb data types to `web-common`
