+++ web-common/src/lib/duckdb-data-types.ts
