+++ web-local/src/lib/components/data-types/FormattedDataType.svelte
@@ -1,7 +1,11 @@
 <script>
   /** provides the formatting for data types */
+  import {
+    INTERVALS,
+    NUMERICS,
+    TIMESTAMPS,
+  } from "@rilldata/web-common/lib/duckdb-data-types";
   import { formatDataType } from "@rilldata/web-common/lib/formatters";
-  import { INTERVALS, NUMERICS, TIMESTAMPS } from "../../duckdb-data-types.ts";
   import Interval from "./Interval.svelte";
   import Number from "./Number.svelte";
   import Timestamp from "./Timestamp.svelte";
