+++ web-local/src/lib/components/viz/histogram/OutlierHistogram.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
+  import { DATA_TYPE_COLORS } from "@rilldata/web-common/lib/duckdb-data-types";
   import { guidGenerator } from "@rilldata/web-common/lib/guid";
   import { fly } from "svelte/transition";
-  import { DATA_TYPE_COLORS } from "../../../duckdb-data-types";
   import HistogramBase from "./HistogramBase.svelte";
 
   export let data;
