+++ web-local/test/data/MetricsExplorer.data.ts
@@ -1,8 +1,8 @@
-import type { TimeSeriesTimeRange } from "@rilldata/web-local/lib/temp/time-control-types";
+import { PreviewRollupInterval } from "@rilldata/web-common/lib/duckdb-data-types";
 import type { MetricsViewRequestFilter } from "@rilldata/web-local/common/rill-developer-service/MetricsViewActions";
-import { PreviewRollupInterval } from "@rilldata/web-local/lib/duckdb-data-types";
-import { getTimeRange } from "../utils/time-series-helpers";
+import type { TimeSeriesTimeRange } from "@rilldata/web-local/lib/temp/time-control-types";
 import type { TimeSeriesMeasureRange } from "../utils/time-series-helpers";
+import { getTimeRange } from "../utils/time-series-helpers";
 
 export interface MetricsExplorerTestDataType {
   title: string;
